Spatially interpolate community-level data

This function will interpolate only numeric variables with names starting with n_, median_, or fraction_. Internally, fraction_ and median_ variables are interpolated non-extensively, but n_ variables are interpolated extensively. Weights at the census block level are used to spatially interpolate different geographies. Block-level total population, total number of homes, or total area from the 2020 Census can be chosen to use for the weights.

Usage

interpolate(from, to, weights = c("pop", "homes", "area"))

Arguments

from

sf object with a neighborhood, census tract, or zcta column and numeric values to be interpolated into target geographies

to

sf object of target geography (must be one of the cincy:: geography objects)

weights

use one of "pop" (population), "homes", or "area" from the 2020 census block estimates to interpolate the values